refactor: move functions to proper folders
