--- contracts/liquidity_hub/bonding-manager/src/distribution/mod.rs
@@ -1,1 +1,0 @@
-pub mod commands;
