+++ contracts/liquidity_hub/bonding-manager/src/lib.rs
@@ -1,8 +1,10 @@
+mod bonding;
 mod commands;
 pub mod contract;
 mod error;
 pub mod helpers;
 mod queries;
+mod rewards;
 pub mod state;
 
 #[cfg(test)]
