+++ contracts/liquidity_hub/bonding-manager/src/rewards/commands.rs
@@ -1,0 +1,296 @@
+use crate::queries::{get_expiring_reward_bucket, query_claimable, query_weight};
+use crate::state::{CONFIG, GLOBAL, LAST_CLAIMED_EPOCH, REWARD_BUCKETS};
+use crate::{helpers, ContractError};
+use cosmwasm_std::{
+    ensure, BankMsg, Coin, CosmosMsg, Decimal, DepsMut, Env, MessageInfo, Order, Response,
+    StdError, StdResult, SubMsg, Uint128, Uint64,
+};
+use white_whale_std::bonding_manager::{GlobalIndex, RewardBucket};
+use white_whale_std::epoch_manager::epoch_manager::Epoch;
+use white_whale_std::pool_network::asset;
+
+pub(crate) fn fill_rewards(
+    deps: DepsMut,
+    env: Env,
+    info: MessageInfo,
+) -> Result<Response, ContractError> {
+    println!("----fill_rewards----");
+
+    // Finding the most recent bucket
+    let upcoming_bucket_id = match REWARD_BUCKETS
+        .keys(deps.storage, None, None, Order::Descending)
+        .next()
+    {
+        Some(bucket_id) => bucket_id?,
+        None => return Err(ContractError::Unauthorized),
+    };
+
+    let config = CONFIG.load(deps.storage)?;
+    let distribution_denom = config.distribution_denom.clone();
+
+    let mut messages: Vec<CosmosMsg> = vec![];
+    let mut submessages: Vec<SubMsg> = vec![];
+    // swap non-whale to whale
+    // Search info funds for LP tokens, LP tokens will contain LP_SYMBOL from lp_common and the string .pair.
+    let mut whale = info
+        .funds
+        .iter()
+        .find(|coin| coin.denom.eq(distribution_denom.as_str()))
+        .unwrap_or(&Coin {
+            denom: distribution_denom.clone(),
+            amount: Uint128::zero(),
+        })
+        .to_owned();
+
+    // coins (not the distribution_denom) that are laying in the contract and have not been swapped before for lack
+    // of swap routes
+    let remanent_coins = deps
+        .querier
+        .query_all_balances(env.contract.address)?
+        .into_iter()
+        .filter(|coin| coin.denom.ne(distribution_denom.as_str()))
+        .collect::<Vec<Coin>>();
+
+    // Each of these helpers will add messages to the messages vector
+    // and may increment the whale Coin above with the result of the swaps
+    helpers::handle_lp_tokens(&remanent_coins, &config, &mut submessages)?;
+    helpers::swap_coins_to_main_token(
+        remanent_coins,
+        &deps,
+        config,
+        &mut whale,
+        &distribution_denom,
+        &mut messages,
+    )?;
+
+    // Add the whale to the funds, the whale figure now should be the result
+    // of all the LP token withdrawals and swaps
+    // Because we are using minimum receive, it is possible the contract can accumulate micro amounts of whale if we get more than what the swap query returned
+    // If this became an issue would could look at replys instead of the query
+    REWARD_BUCKETS.update(deps.storage, upcoming_bucket_id, |bucket| -> StdResult<_> {
+        let mut bucket = bucket.unwrap_or_default();
+        bucket.available = asset::aggregate_coins(bucket.available, vec![whale.clone()])?;
+        bucket.total = asset::aggregate_coins(bucket.total, vec![whale.clone()])?;
+        Ok(bucket)
+    })?;
+    Ok(Response::default()
+        .add_messages(messages)
+        .add_submessages(submessages)
+        .add_attributes(vec![("action", "fill_rewards".to_string())]))
+}
+
+pub(crate) fn on_epoch_created(
+    deps: DepsMut,
+    info: MessageInfo,
+    current_epoch: Epoch,
+) -> Result<Response, ContractError> {
+    cw_utils::nonpayable(&info)?;
+
+    println!("----on_epoch_created----");
+    println!("EpochChangedHook: {:?}", current_epoch);
+    // A new epoch has been created, update rewards bucket and forward the expiring bucket
+    let config = CONFIG.load(deps.storage)?;
+    ensure!(
+        info.sender == config.epoch_manager_addr,
+        ContractError::Unauthorized
+    );
+
+    let global = GLOBAL.may_load(deps.storage)?;
+    // This happens only on the very first epoch where Global has not been initialised yet
+    if global.is_none() {
+        let initial_global_index = GlobalIndex {
+            updated_last: current_epoch.id,
+            ..Default::default()
+        };
+        GLOBAL.save(deps.storage, &initial_global_index)?;
+        REWARD_BUCKETS.save(
+            deps.storage,
+            current_epoch.id,
+            &RewardBucket {
+                id: current_epoch.id,
+                epoch_start_time: current_epoch.start_time,
+                global_index: initial_global_index,
+                ..RewardBucket::default()
+            },
+        )?;
+    }
+
+    // Update the global index epoch id field
+    let mut global = GLOBAL.load(deps.storage)?;
+    global.epoch_id = current_epoch.id;
+
+    // update the global index for the current bucket, take the current snapshot of the global index
+    REWARD_BUCKETS.update(
+        deps.storage,
+        current_epoch.id,
+        |reward_bucket| -> StdResult<_> {
+            let mut reward_bucket = reward_bucket.unwrap_or_default();
+            reward_bucket.global_index = global;
+            Ok(reward_bucket)
+        },
+    )?;
+
+    // forward fees from the expiring bucket to the new one.
+    let mut expiring_reward_bucket = get_expiring_reward_bucket(deps.as_ref())?;
+    if let Some(expiring_bucket) = expiring_reward_bucket.as_mut() {
+        // Load all the available assets from the expiring bucket
+        let amount_to_be_forwarded = REWARD_BUCKETS
+            .load(deps.storage, expiring_bucket.id)?
+            .available;
+        REWARD_BUCKETS.update(deps.storage, current_epoch.id, |bucket| -> StdResult<_> {
+            let mut bucket = bucket.unwrap_or_default();
+            bucket.available =
+                asset::aggregate_coins(bucket.available, amount_to_be_forwarded.clone())?;
+            bucket.total = asset::aggregate_coins(bucket.total, amount_to_be_forwarded)?;
+
+            Ok(bucket)
+        })?;
+        // Set the available assets for the expiring bucket to an empty vec now that they have been
+        // forwarded
+        REWARD_BUCKETS.update(deps.storage, expiring_bucket.id, |bucket| -> StdResult<_> {
+            let mut bucket = bucket.unwrap_or_default();
+            bucket.available = vec![];
+            Ok(bucket)
+        })?;
+    }
+
+    // Create a new bucket for the rewards flowing from this time on, i.e. to be distributed in
+    // the next epoch. Also, forwards the expiring bucket (only 21 bucket are live at a given moment)
+    let next_epoch_id = Uint64::new(current_epoch.id)
+        .checked_add(Uint64::one())?
+        .u64();
+    REWARD_BUCKETS.save(
+        deps.storage,
+        next_epoch_id,
+        &RewardBucket {
+            id: next_epoch_id,
+            epoch_start_time: current_epoch.start_time.plus_days(1),
+            // this global index is to be updated the next time this hook is called, as this future epoch
+            // will become the current one
+            global_index: Default::default(),
+            ..RewardBucket::default()
+        },
+    )?;
+
+    Ok(Response::default().add_attributes(vec![("action", "epoch_changed_hook".to_string())]))
+}
+
+/// Claims pending rewards for the sender.
+pub fn claim(deps: DepsMut, info: MessageInfo) -> Result<Response, ContractError> {
+    let claimable_reward_buckets_for_user =
+        query_claimable(deps.as_ref(), Some(info.sender.to_string()))?.reward_buckets;
+    ensure!(
+        !claimable_reward_buckets_for_user.is_empty(),
+        ContractError::NothingToClaim
+    );
+
+    let mut claimable_rewards = vec![];
+    let mut attributes = vec![];
+    for mut reward_bucket in claimable_reward_buckets_for_user.clone() {
+        let bonding_weight_response_for_epoch = query_weight(
+            deps.as_ref(),
+            reward_bucket.id,
+            info.sender.to_string(),
+            Some(reward_bucket.global_index.clone()),
+        )?;
+
+        // if the user has no share in the bucket, skip it
+        if bonding_weight_response_for_epoch.share.is_zero() {
+            continue;
+        };
+
+        // sanity check
+        ensure!(
+            bonding_weight_response_for_epoch.share <= Decimal::percent(100u64),
+            ContractError::InvalidShare
+        );
+
+        for reward in reward_bucket.total.iter() {
+            let user_reward = reward.amount * bonding_weight_response_for_epoch.share;
+
+            // make sure the reward is sound
+            let reward_validation: Result<(), StdError> = reward_bucket
+                .available
+                .iter()
+                .find(|available_fee| available_fee.denom == reward.denom)
+                .map(|available_fee| {
+                    if user_reward > available_fee.amount {
+                        attributes.push((
+                            "error",
+                            ContractError::InvalidReward {
+                                reward: user_reward,
+                                available: available_fee.amount,
+                            }
+                            .to_string(),
+                        ));
+                    }
+                    Ok(())
+                })
+                .ok_or(StdError::generic_err("Invalid fee"))?;
+
+            // if the reward is invalid, skip the bucket
+            match reward_validation {
+                Ok(_) => {}
+                Err(_) => continue,
+            }
+
+            let denom = &reward.denom;
+            // add the reward
+            claimable_rewards = asset::aggregate_coins(
+                claimable_rewards,
+                vec![Coin {
+                    denom: denom.to_string(),
+                    amount: user_reward,
+                }],
+            )?;
+
+            // modify the bucket to reflect the new available and claimed amount
+            for available_fee in reward_bucket.available.iter_mut() {
+                if available_fee.denom == reward.denom {
+                    available_fee.amount = available_fee.amount.saturating_sub(user_reward);
+                }
+            }
+
+            if reward_bucket.claimed.is_empty() {
+                reward_bucket.claimed = vec![Coin {
+                    denom: denom.to_string(),
+                    amount: user_reward,
+                }];
+            } else {
+                for claimed_reward in reward_bucket.claimed.iter_mut() {
+                    if claimed_reward.denom == reward.denom {
+                        claimed_reward.amount = claimed_reward.amount.checked_add(user_reward)?;
+                    }
+
+                    // sanity check, should never happen
+                    for total_reward in reward_bucket.total.iter() {
+                        if total_reward.denom == claimed_reward.denom {
+                            ensure!(
+                                claimed_reward.amount <= total_reward.amount,
+                                ContractError::InvalidShare
+                            );
+                        }
+                    }
+                }
+            }
+
+            REWARD_BUCKETS.save(deps.storage, reward_bucket.id, &reward_bucket)?;
+        }
+    }
+
+    // update the last claimed epoch for the user. it's in the first bucket on the list since it's sorted
+    // in descending order
+    LAST_CLAIMED_EPOCH.save(
+        deps.storage,
+        &info.sender,
+        &claimable_reward_buckets_for_user[0].id,
+    )?;
+
+    Ok(Response::default()
+        .add_attributes(vec![("action", "claim".to_string())])
+        .add_attributes(attributes)
+        .add_message(CosmosMsg::Bank(BankMsg::Send {
+            to_address: info.sender.to_string(),
+            amount: claimable_rewards,
+        })))
+}
